Rename `OutboxRelay` to `RelayMessagesThroughConsumer`
This makes sure the class name conveys its intent, whilst keeping
backwards compatibility.
